Refactors HTTP method location
Moves the HTTP method enum to a shared location to be accessible by multiple modules. This avoids duplication and ensures consistency across the application.
